requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.16.4
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
attrs==24.3.0
Authlib==1.6.1
blinker==1.9.0
bottleneck
brotlicffi==1.0.9.2
cachetools==5.5.1
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.8
colorama==0.4.6
cryptography==45.0.6
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.7.0
easyocr
email_validator
faiss-cpu
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
filelock==3.18.0
filetype==1.2.0
frozenlist==1.7.0
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.43
#google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-vision==3.4.5
google-cloud-storage
google-generativeai
googleapis-common-protos==1.70.0
# google-cloud-sql-connector[pg8000]
greenlet==3.2.3
grpcio==1.74.0
grpcio-status==1.62.3
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
idna==3.7
imageio
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2023.7.1
langchain==0.3.27
langchain-community
langchain-core==0.3.72
langchain-google-genai
langchain-openai==0.3.28
langchain-text-splitters==0.3.9
langsmith==0.4.11
lazy_loader==0.4
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mkl_fft
mkl_random
mkl-service
mpmath==1.3.0
multidict==6.6.3
mypy_extensions==1.1.0
narwhals==1.31.0
nest-asyncio==1.6.0
networkx==3.5
ninja==1.11.1.4
numexpr
numpy
openai==1.98.0
opencv-python-headless
orjson==3.11.1
packaging==25.0
pandas==2.3.1
pillow==11.1.0
pip==25.1
propcache==0.3.2
proto-plus==1.26.1
protobuf==4.25.8
psutil==7.0.0
#psycopg2
psycopg2-binary==2.9.10
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pycparser==2.21
pydantic==2.11.7
pydantic_core==2.33.2
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
Pygments==2.19.2
PyJWT==2.10.1
PyMuPDF==1.26.3
pymupdf4llm==0.0.27
pyparsing==3.2.3
PySocks==1.7.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.30.2
regex==2025.7.34
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.1.0
rich-toolkit==0.14.9
rignore==0.6.4
rpds-py==0.22.3
rsa==4.9.1
scikit-image==0.25.2
scipy==1.16.1
sentry-sdk==2.34.1
setuptools==72.1.0
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.42
sqlmodel==0.0.24
starlette==0.47.2
streamlit==1.47.0
streamlit-cookies-manager==0.2.0
sympy==1.14.0
tenacity==9.0.0
tifffile==2025.6.11
tiktoken==0.9.0
toml==0.10.2
torch==2.7.1
torchvision==0.22.1
tornado==6.5.1
tqdm==4.67.1
typer==0.16.0
typing_extensions==4.12.2
typing-inspect==0.9.0
typing-inspection==0.4.1
tzdata==2025.2
ujson==5.10.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
watchdog==4.0.2
watchfiles==1.1.0
websockets==15.0.1
wheel==0.45.1
win-inet-pton==1.1.0
yarl==1.20.1
zstandard==0.23.0
